Working build on Windows
In addition to the usual JDK + Maven, this depends on Git (with bash), CMake, and Visual Studio. Bash provided with Git is used to execute the build script, CMake is used to invoke the build system rather than calling Make directly.

Some odd edges like the strange return code from exec, and hard coding the Win64 build system. No detection of 32bit vs 64bit, and instead the script tries to force things to be 64bit, which necessitates specifying the generator. That can perhaps be smoothed over further with the Windows profile in pom.xml.