Set architecture on Windows instead of Generator
This seems like a more robust way of specifying the build, since we are
not assuming specific versions of Visual Studio. The behavior in this
case seems to be for CMake to attempt to the detect the latest version
of Visual Studio it can use to build the project.

CMAKE_BUILD_TYPE may be unused now - CMAKE warns it is set but not used.