Upgrade to v4.0.0-rc2 (#91)
Upgrades H3-Java to H3 core library v4.0.0 release candidate. Since this is not the final v4 release, H3-Java with v4 is also considered pre-release.

A new class, `H3CoreV3`, is added to act as a compatibility layer during upgrades. Because it is implemented using the v4 library underneath, care was not taken to make exceptions the same across both interfaces - only the function names are the same. I think this makes H3CoreV3 of limited use, so I expect to remove it in an update before or soon after releasing h3-java v4.0.0.